requirements.txt
fastapi
keras
numpy
pandas
Pillow
scikit_learn
tensorflow
uvicorn
numpy
typing_extensions
python-multipart
